+++ checkstyle.xml
@@ -45,7 +45,7 @@
   <!-- See http://checkstyle.sf.net/config_misc.html#Translation -->
   <module name="Translation"/>
 
-  <!-- module name="FileLength"/ -->
+  <module name="FileLength"/>
   <module name="FileTabCharacter"/>
 
   <module name="TreeWalker">
@@ -96,7 +96,6 @@
     <!-- Checks for Size Violations.                    -->
     <!-- See http://checkstyle.sf.net/config_sizes.html -->
     <module name="LineLength"/>
-    <!-- module name="MethodLength"/ -->
     <module name="ParameterNumber"/>
 
     <!-- Checks for whitespace                               -->
